move line executor to executor file
